Major Overhaul
Remade a bunch of different stuff. Added more functionality to the Hardware Driver, as well as began the consolidation of most of the code into one file so class instances could be shared. Finally started the re writing of the Fuel injector scripts as well as added some much needed commenting throughout the code.

Am now going to sleep but happy with progress.